fixed retrieval of challenger modules in challenge system
Previously tests were failing (as noted in the preceding commit message)
because we were erroneously reusing _determine_challenger, which did not
provide the proper level of encapsulation.

We've now extracted a separate method for accessing/importing challenger
modules, resolving the earlier confusion.

NB: Exceptions due to failed imports are ignored (swallowed) when
listing the available auth systems (in contrast to requesting a
particular challenger via the HTTP API). This might not be desirable -
it is primarily required to make the current tests run smoothly, but
should not normally occur.